integrate pw and multisig-v2 into dev spec (#280)
* integrate pw and multisig-v2 into dev spec

* fix path

* refactor private script lib pos
